backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.7.0
python-pptx>=1.0.0
github-copilot-sdk>=0.1.0
azure-ai-projects>=1.0.0b1
azure-identity>=1.17.0
